BACKEND/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-froala-editor==4.3.0
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
